api/requirements.txt
django
djangorestframework
celery
pandas
redis
requests
